+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Improvements
 * Added support for PyNWB 3.0 [#557](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/557)
 * Added support for Python 3.13 [#564](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/564)
+* Added new check function `check_ascending_spike_times` with `nelems` parameter to verify spike times are in ascending order [PR #405](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/405)
 
 
 # v0.6.2